LayoutTests/platform/win/http/tests/misc/webtiming-two-redirects-expected.txt

This test checks that Web Timing reports redirectCount correctly and has reasonable values for redirectStart and redirectCount.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


FAIL timing.redirectStart should be >= timing.navigationStart. Was 0 (of type number).
PASS timing.redirectEnd is >= timing.redirectStart
FAIL navigation.redirectCount should be 2. Was 0.
PASS successfullyParsed is true

TEST COMPLETE
